fix bugs in parsing glyf table
Moving the entire let form into the vec type form had introduced a bug that the
index counter resets instead of counts. This also fixes the bug that the index
counter wasn’t being reset each time the spec is used.

The loca map is now cached in a local memoized accessor.