fix alignment and loca table caching bugs in glyf data parsing
Glyf data alignment:

For some fonts glyf data sections are aligned and word boundaries and for
others on double word. Instead fill to the known boundary from the loca table,
which saves having to worry about alignment on a font by font basis.

Loca table caching:

Remove the accessor closure and instead sets a local var to hold the loca table
for subsequent accesses in building the glyph data vector.