add elisp implementation of the fontdue raster
It should be noted that the fontdue author has this to say:
“Notice to anyone that wants to repurpose the raster for your library: Please
don't reuse this raster. Fontdue's raster is very unsafe, with nuanced
invariants that need to be accounted for. Fontdue sanitizes the input that the
raster will consume to ensure it is safe. Please be aware of this.”

The elisp implementation faithfully reproduces fontdue’s output for multiple
fonts at various pixel sizes. That said, the plan is to have multiple raster
backends that use different strategies, so this is just a start.