add a helper fn to concatenate single line horizontal layouts
Note this only works for single line horizontal layouts. As such, it's not
generally useful and should be obsoleted prior to API finalization.

This is a workaround for layouts containing segments at different pixel
sizes. In that case, y coords for subsequent segment positions are different
from those if the same segment were in a layout by itself.