update layout-current-pos during layout-finalize
This has layout finalize update the layout-current-pos according to
layout-settings-x in addition to updating the glyph-position-x values. This
just means that current-pos will have the current-pos of the last line plus the
layout-settings-x. If it turns out this is somehow harmful, alternatives could
be for callers to add their settings-x to the current-pos if desired, or to add
a field to layout to hold all actual line widths after finalize.

For my purposes this means I can remove fontsloth-layout-concat-layouts, which
is doing unneeded extra work and is not generally useful.