add bindat specs and fns for GPOS horizontal pair positioning data
aka kerning

This also tests that the indexing doesn’t include invalid entries. Currently it
only produces an index for the default script language. It appears that many
fonts will have multiple langs point to the same kern feature tables anyway, so
the it may be good to only index additional lang kern features if they differ
from the default.