Include cmap format0, format6, and format12 in char-to-glyph map
This is in addition to format4 which was already included.

Cmap subtables are now read starting at offsets specified in the encodings
array. Subtables are de-duplicated in the case of multiple encodings pointing
to the same subtable offset.

The char-to-glyph map is now a hash-table instead of an alist. This is
reflected in fontsloth--load-font which iterates char-to-glyph values to decide
with font glyphs to outline.