Fully handle CFF1 for SID compatible encodings
New dependency: stream 2.2.5

This should outline glyphs for most free OTF fonts I’ve encountered so far.

With regard to the entirety of CFF1/CFF2, this code is incomplete in at least
the following ways:
+ No outlines are produced for CFF2 tables.
+ CID metadata isn’t parsed which means that local subroutines for CID
  encodings are not interpreted (i.e. outlines are incomplete).
+ Sid->gid mapping is currently only supported for format1 charsets (meaning
  endchar base and accent char lookups will fail for format0 and format2
  charsets (these are anecdotally much less common than format1, but it’s
  still a must to add support).

Notes about the code itself:

It is modeled after the Rust ttf-parser CFF outliner.

The bulk of code is a postscript language interpreter. It may be good to build
a more idiomatic, standalone elisp postscript interpreter. There are also a
number of potential performance improvements for the interpreter such as
effectively compiling subroutines on first encounter.